LayoutTests/fast/webgpu/loop-break-continuing-no-switch-expected.txt

CONSOLE MESSAGE: 1000
CONSOLE MESSAGE: no validation error
